- Move the project from src folder to the root folder - Use Q_UNUSED for unused variables - Change "class ElementInfo;" to "struct ElementInfo;" in reader.cpp - Move the headers to qtcsv folder
