+++ .gitignore
@@ -33,3 +33,4 @@ build-*
 # Project
 *tst_*
 
+libs/
