+++ include/qtcsv/abstractdata.h
