+++ include/qtcsv/qtcsv_global.h
