+++ include/qtcsv/stringdata.h
