+++ include/qtcsv/variantdata.h
