+++ sources/contentiterator.h
