+++ sources/filechecker.h
