+++ sources/symbols.h
